backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
anyio>=4.4.0
python-dotenv>=1.0.0
